// src/store/types/Health/Form/THealthFormResponse.tsx
export type THealthFormResponse = {
  sum_insured: number;
  quote_no: string;
  module: string;
  insurance_type: string;
  term: number;
  cover: string;
  state: string;
  city: string;
  child: number;
  adult: number;
  sort_by?: string;
  mobile?: number;
  pincode?: string;
  utm_medium?: string;
  utm_source?: string;
  Number_of_Plans_Listed: number;
  Number_of_Brands_Listed: number;
};